mqtt/pyproject.toml
[build-system]
requires = ["hatchling >= 1.26.3", "pyspark >= 4.0.0.dev2", "paho-mqtt >= 2.1.0"]
build-backend = "hatchling.build"

[project]
name = "python_datasource_connectors"
version = "0.0.2"
authors = [
    { name="Jeffery Annor", email="jeffery.annor@databricks.com" },
    { name="Gaston Guitart", email="gaston.guitart@databricks.com" },
]
description = "A Collection of Custom Data Source Connectors for PySpark"
requires-python = ">=3.11"
dependencies = [
    "pyspark >= 4.0.0.dev2",
    "paho-mqtt >= 2.1.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]
doc = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  "src"
]

readme = "README.md"
license = "EPL-2.0"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/jefferyann-db/python-datasource-connectors/mqtt"
Issues = "https://github.com/jefferyann-db/python-datasource-connectors/issues"